Refactor fetcher package structure
This commit fixes the fetcher package structure to include the package name `fetcher` before any subpackage (for example: `com.xkcddatahub.fetcher.domain` instead of `com.xkcddatahub.domain`)
