--- fetcher/src/main/kotlin/com/xkcddatahub/application/ports/WebComicsPort.kt
@@ -1,9 +1,0 @@
-package com.xkcddatahub.application.ports
-
-import com.xkcddatahub.domain.entity.WebComics
-
-interface WebComicsPort {
-    suspend fun save(comics: WebComics): Boolean
-
-    suspend fun getLatestStoredComicId(): Int
-}
