--- fetcher/src/main/kotlin/com/xkcddatahub/application/ports/XkcdClientPort.kt
@@ -1,9 +1,0 @@
-package com.xkcddatahub.application.ports
-
-import com.xkcddatahub.domain.entity.WebComics
-
-interface XkcdClientPort {
-    suspend fun getLatestComicId(): Int
-
-    suspend fun getComicsById(id: Int): WebComics
-}
