--- fetcher/src/main/kotlin/com/xkcddatahub/bootstrap/Application.kt
@@ -1,27 +1,0 @@
-package com.xkcddatahub.bootstrap
-
-import com.xkcddatahub.application.usecases.GetAllXkcdComics
-import com.xkcddatahub.bootstrap.di.allModules
-import io.ktor.server.application.Application
-import io.ktor.server.application.install
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import org.koin.ktor.ext.inject
-import org.koin.ktor.plugin.Koin
-
-fun main(args: Array<String>) {
-    io.ktor.server.netty.EngineMain.main(args)
-}
-
-fun Application.module() {
-    install(Koin) {
-        modules(allModules)
-    }
-
-    DatabaseInitializer(environment.config).init()
-
-    launch(Dispatchers.IO) {
-        val getAllXkcdComics by inject<GetAllXkcdComics>()
-        getAllXkcdComics()
-    }
-}
