--- fetcher/src/main/kotlin/com/xkcddatahub/bootstrap/di/ApplicationModule.kt
@@ -1,9 +1,0 @@
-package com.xkcddatahub.bootstrap.di
-
-import com.xkcddatahub.application.usecases.GetAllXkcdComics
-import org.koin.dsl.module
-
-val appModule =
-    module {
-        single { GetAllXkcdComics(get(), get()) }
-    }
