--- fetcher/src/main/kotlin/com/xkcddatahub/bootstrap/di/KoinModules.kt
@@ -1,8 +1,0 @@
-package com.xkcddatahub.bootstrap.di
-
-val allModules =
-    listOf(
-        networkModule,
-        adapterModule,
-        appModule,
-    )
