--- fetcher/src/main/kotlin/com/xkcddatahub/bootstrap/di/NetworkModule.kt
@@ -1,18 +1,0 @@
-package com.xkcddatahub.bootstrap.di
-
-import io.ktor.client.HttpClient
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.serialization.json.Json
-import org.koin.dsl.module
-
-val networkModule =
-    module {
-        single {
-            HttpClient {
-                install(ContentNegotiation) {
-                    json(Json { ignoreUnknownKeys = true })
-                }
-            }
-        }
-    }
