+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/adapters/output/database/postgres/mappers/DatabaseWebComicsMapper.kt
@@ -1,0 +1,21 @@
+package com.xkcddatahub.fetcher.adapters.output.database.postgres.mappers
+
+import com.xkcddatahub.fetcher.adapters.output.database.postgres.data.DatabaseWebComics
+import com.xkcddatahub.fetcher.entity.WebComics
+
+object DatabaseWebComicsMapper {
+    fun WebComics.toData() =
+        DatabaseWebComics(
+            id = id,
+            year = year,
+            month = month,
+            day = day,
+            title = title,
+            safeTitle = safeTitle,
+            transcript = transcript,
+            alternativeText = alternativeText,
+            imageUrl = imageUrl,
+            news = news,
+            link = link,
+        )
+}
