+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/application/ports/WebComicsPort.kt
@@ -1,0 +1,9 @@
+package com.xkcddatahub.fetcher.application.ports
+
+import com.xkcddatahub.fetcher.entity.WebComics
+
+interface WebComicsPort {
+    suspend fun save(comics: WebComics): Boolean
+
+    suspend fun getLatestStoredComicId(): Int
+}
