+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/application/ports/XkcdClientPort.kt
@@ -1,0 +1,9 @@
+package com.xkcddatahub.fetcher.application.ports
+
+import com.xkcddatahub.fetcher.entity.WebComics
+
+interface XkcdClientPort {
+    suspend fun getLatestComicId(): Int
+
+    suspend fun getComicsById(id: Int): WebComics
+}
