+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/bootstrap/di/ApplicationModule.kt
@@ -1,0 +1,9 @@
+package com.xkcddatahub.fetcher.bootstrap.di
+
+import com.xkcddatahub.fetcher.application.usecases.GetAllXkcdComics
+import org.koin.dsl.module
+
+val appModule =
+    module {
+        single { GetAllXkcdComics(get(), get()) }
+    }
