+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/bootstrap/di/KoinModules.kt
@@ -1,0 +1,8 @@
+package com.xkcddatahub.fetcher.bootstrap.di
+
+val allModules =
+    listOf(
+        networkModule,
+        adapterModule,
+        appModule,
+    )
