+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/bootstrap/di/NetworkModule.kt
@@ -1,0 +1,18 @@
+package com.xkcddatahub.fetcher.bootstrap.di
+
+import io.ktor.client.HttpClient
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+import org.koin.dsl.module
+
+val networkModule =
+    module {
+        single {
+            HttpClient {
+                install(ContentNegotiation) {
+                    json(Json { ignoreUnknownKeys = true })
+                }
+            }
+        }
+    }
