+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/utilities/DateTimeUtils.kt
@@ -1,0 +1,12 @@
+package com.xkcddatahub.fetcher.utilities
+
+import java.time.Instant
+import java.time.LocalDateTime
+import java.time.ZoneOffset
+
+object DateTimeUtils {
+    fun currentUtc(): LocalDateTime =
+        Instant.now()
+            .atOffset(ZoneOffset.UTC)
+            .toLocalDateTime()
+}
