--- fetcher/src/main/kotlin/com/xkcddatahub/utilities/DateTimeUtils.kt
@@ -1,12 +1,0 @@
-package com.xkcddatahub.utilities
-
-import java.time.Instant
-import java.time.LocalDateTime
-import java.time.ZoneOffset
-
-object DateTimeUtils {
-    fun currentUtc(): LocalDateTime =
-        Instant.now()
-            .atOffset(ZoneOffset.UTC)
-            .toLocalDateTime()
-}
