--- fetcher/src/test/kotlin/com/xkcddatahub/adapters/output/database/postgres/WebComicsPostgresAdapterTest.kt
@@ -1,5 +1,0 @@
-package com.xkcddatahub.adapters.output.database.postgres
-
-class WebComicsPostgresAdapterTest {
-    // TODO: create test with test containers
-}
