+++ fetcher/src/test/kotlin/com/xkcddatahub/fetcher/output/database/postgres/WebComicsPostgresAdapterTest.kt
@@ -1,0 +1,5 @@
+package com.xkcddatahub.fetcher.output.database.postgres
+
+class WebComicsPostgresAdapterTest {
+    // TODO: create test with test containers
+}
